+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Support a package:native type option for specifying native type
 - Add a preferences menu for Windows (since it does not have one out of the box)
 - Fix app scaffold on Windows by having it generate jeweler gem first (to have gemspec for jar-dependencies)
+- Fix girb for Windows
 
 ### 0.6.0